Review of Version 1
Version 1 review and submittal
